Add mssql.proposed.d.ts
